WIP - One step back, adding things back to programming view so we can be waise when creating reusable components'
