+++ src/components/features/filters/components/SearchFilter.vue
@@ -2,7 +2,7 @@
 // import { isEqual } from "lodash-es";
 import { computed, ref } from "vue";
 import { toRaw } from "vue";
-import { IconFilter, IconClose, IconSearch } from "@/components/common/icons";
+import { IconClose,  } from "@/components/common/icons";
 import TwContainer from "@/components/layout/TwContainer.vue";
 import AccordionGroup from "@/components/base/accordion/AccordionGroup.vue";
 import ComboboxComponent from "@/components/common/forms/components/ComboboxComponent.vue";
@@ -13,222 +13,16 @@ import SelectComponent from "@/components/common/forms/components/SelectComponen
 import { generateTimeOptions } from "@/components/features/filters/composables/useTimeOptions";
 import { useQuery } from "@tanstack/vue-query";
 import { cleanObject } from "@/utils/helpers/objectHelpers";
+import { collection, showcases } from "@/lib/fakeData"
 
-const collection = [
-  { value: "Joachim Trier", label: "Joachim Trier" },
-  { value: "Lynne Ramsay", label: "Lynne Ramsay" },
-  {
-    value: "Isabel Joffily e Pedro Rossi",
-    label: "Isabel Joffily e Pedro Rossi",
-  },
-  { value: "Julia Ducournau", label: "Julia Ducournau" },
-  {
-    value: "Marcio Reolon e Filipe Matzembacher",
-    label: "Marcio Reolon e Filipe Matzembacher",
-  },
-  { value: "Lloyd Lee Choi", label: "Lloyd Lee Choi" },
-  {
-    value: "Bruno Jorge, Mariana Oliva e Renato Terra",
-    label: "Bruno Jorge, Mariana Oliva e Renato Terra",
-  },
-  { value: "Françoise Ferraton", label: "Françoise Ferraton" },
-  { value: "Natesh Hegde", label: "Natesh Hegde" },
-  {
-    value: "Delphine Coulin e Muriel Coulin",
-    label: "Delphine Coulin e Muriel Coulin",
-  },
-  { value: "Maya Da-Rin", label: "Maya Da-Rin" },
-  { value: "Luciano Zito", label: "Luciano Zito" },
-  { value: "Anna Cornudella", label: "Anna Cornudella" },
-  { value: "Daniel Bandeira", label: "Daniel Bandeira" },
-  { value: "Mehrnoush Alia", label: "Mehrnoush Alia" },
-  { value: "Adriana L. Dutra", label: "Adriana L. Dutra" },
-  { value: "João Rosas", label: "João Rosas" },
-  {
-    value: "Vincent Carelli, Tatiana Almeida e Ernesto de Carvalho",
-    label: "Vincent Carelli, Tatiana Almeida e Ernesto de Carvalho",
-  },
-  { value: "Vibeke Løkkeberg", label: "Vibeke Løkkeberg" },
-  { value: "Vivian Qu", label: "Vivian Qu" },
-  { value: "Tereza Nvotová", label: "Tereza Nvotová" },
-  {
-    value: "Luiz Bolognesi, Jean Cullen de Moura e Marcelo Fernandes de Moura",
-    label: "Luiz Bolognesi, Jean Cullen de Moura e Marcelo Fernandes de Moura",
-  },
-  { value: "Luca Guadagnino", label: "Luca Guadagnino" },
-  {
-    value: "Aurélien Vernhes-Lermusiaux",
-    label: "Aurélien Vernhes-Lermusiaux",
-  },
-  {
-    value: "Eleanor Coppola, Fax Bahr e George Hickenlooper",
-    label: "Eleanor Coppola, Fax Bahr e George Hickenlooper",
-  },
-  { value: "Ido Fluk", label: "Ido Fluk" },
-  {
-    value: "Carlos Fausto, Takumã Kuikuro e Leonardo Sette",
-    label: "Carlos Fausto, Takumã Kuikuro e Leonardo Sette",
-  },
-  { value: "Emilie Blichfeldt", label: "Emilie Blichfeldt" },
-  { value: "Ondřej Provazník", label: "Ondřej Provazník" },
-  { value: "Estêvão Ciavatta", label: "Estêvão Ciavatta" },
-  { value: "Ebs Burnough", label: "Ebs Burnough" },
-  { value: " Sang-il Lee", label: " Sang-il Lee" },
-  { value: "Pedro de Filippis", label: "Pedro de Filippis" },
-  { value: "Hong Sang-soo", label: "Hong Sang-soo" },
-  {
-    value: "Philip Delaquis e Barbara Miller ",
-    label: "Philip Delaquis e Barbara Miller ",
-  },
-  { value: "Alexe Poukine", label: "Alexe Poukine" },
-  {
-    value: "Tati Franklin  e Suellen Vasconcelos ",
-    label: "Tati Franklin  e Suellen Vasconcelos ",
-  },
-  { value: "Radu Jude", label: "Radu Jude" },
-  { value: "Silvina Schnicer", label: "Silvina Schnicer" },
-  { value: "Pauline Loquès", label: "Pauline Loquès" },
-  { value: "Bruno Murtinho", label: "Bruno Murtinho" },
-  { value: "Mercedes Bryce Morgan", label: "Mercedes Bryce Morgan" },
-  { value: "Jean-François Ravagnan ", label: "Jean-François Ravagnan " },
-  {
-    value: "Marco Altberg e Tainá De Luccas",
-    label: "Marco Altberg e Tainá De Luccas",
-  },
-  {
-    value:
-      "Toby Schmutzler, Kevin Schmutzler, Apuu Mourine Munyes e Vallentine Chelluget",
-    label:
-      "Toby Schmutzler, Kevin Schmutzler, Apuu Mourine Munyes e Vallentine Chelluget",
-  },
-  { value: "Denis Villeneuve", label: "Denis Villeneuve" },
-  { value: "Eduardo Ades", label: "Eduardo Ades" },
-  { value: "Toby Perl Freilich", label: "Toby Perl Freilich" },
-  { value: "Fanny Ovesen", label: "Fanny Ovesen" },
-  { value: "Mary Bronstein", label: "Mary Bronstein" },
-  { value: "Andrea Tonacci", label: "Andrea Tonacci" },
-  { value: "Dolores Fonzi", label: "Dolores Fonzi" },
-  { value: "Mehmet Akif Büyükatalay", label: "Mehmet Akif Büyükatalay" },
-  { value: "Ethan Coen", label: "Ethan Coen" },
-  { value: "Sammy Baloji", label: "Sammy Baloji" },
-  {
-    value: "Renée Nader Messora e João Salaviza",
-    label: "Renée Nader Messora e João Salaviza",
-  },
-  { value: "Mohamed Rashad", label: "Mohamed Rashad" },
-  { value: "Camila Freitas", label: "Camila Freitas" },
-  { value: "Billy Shebar", label: "Billy Shebar" },
-  { value: "Gemma Blasco", label: "Gemma Blasco" },
-  { value: "Rebecca Zlotowski", label: "Rebecca Zlotowski" },
-  { value: "Jorge Furtado", label: "Jorge Furtado" },
-  { value: "Laura Casabe", label: "Laura Casabe" },
-  { value: "Michael Tyburski", label: "Michael Tyburski" },
-  { value: "Teresa Villaverde", label: "Teresa Villaverde" },
-  {
-    value: "Lucy Walker, Karen Harley e João Jardim",
-    label: "Lucy Walker, Karen Harley e João Jardim",
-  },
-  {
-    value: "Chelsea Greene, Rob Grobman e Edivan Guajajara",
-    label: "Chelsea Greene, Rob Grobman e Edivan Guajajara",
-  },
-  {
-    value: "Alessio Rigo de Righi e Matteo Zoppis",
-    label: "Alessio Rigo de Righi e Matteo Zoppis",
-  },
-  { value: "Pia Marais", label: "Pia Marais" },
-  {
-    value: "Vivian Ostrovsky e Ruti Gadish",
-    label: "Vivian Ostrovsky e Ruti Gadish",
-  },
-  {
-    value: "Charlotte Devillers e Arnaud Dufeys",
-    label: "Charlotte Devillers e Arnaud Dufeys",
-  },
-  { value: "Agnieszka Holland", label: "Agnieszka Holland" },
-  { value: "Abderrahmane Sissako", label: "Abderrahmane Sissako" },
-  { value: "Iván Fund", label: "Iván Fund" },
-  { value: "Čejen Černić Čanak", label: "Čejen Černić Čanak" },
-  { value: "Paolo Sorrentino", label: "Paolo Sorrentino" },
-  { value: "Matteo Garrone", label: "Matteo Garrone" },
-  { value: "Chase Joynt", label: "Chase Joynt" },
-  { value: "Alireza Khatami", label: "Alireza Khatami" },
-  { value: "Anna Cazenave Cambet", label: "Anna Cazenave Cambet" },
-  {
-    value: "Mariano Cohn e Gastón Duprat",
-    label: "Mariano Cohn e Gastón Duprat",
-  },
-  { value: "Morad Mostafa", label: "Morad Mostafa" },
-  { value: "Diego Céspedes", label: "Diego Céspedes" },
-  { value: "Sergei Loznitsa", label: "Sergei Loznitsa" },
-  { value: "Peter Kerekes", label: "Peter Kerekes" },
-  { value: "Claire Simon", label: "Claire Simon" },
-  { value: "Avelina Prat", label: "Avelina Prat" },
-  { value: "Kristen Stewart", label: "Kristen Stewart" },
-  { value: "Ernesto Martínez Bucio", label: "Ernesto Martínez Bucio" },
-  { value: "Jaume Claret Muxart", label: "Jaume Claret Muxart" },
-  { value: "Nadav Lapid", label: "Nadav Lapid" },
-  { value: "Maria Eriksson-Hecht", label: "Maria Eriksson-Hecht" },
-  { value: "Allan Ribeiro", label: "Allan Ribeiro" },
-  { value: "Yolanda Centeno", label: "Yolanda Centeno" },
-  { value: "Kelly Reichardt", label: "Kelly Reichardt" },
-  { value: "Davi Pretto", label: "Davi Pretto" },
-  { value: "Ratchapoom Boonbunchachoke", label: "Ratchapoom Boonbunchachoke" },
-  { value: "Tina Romero", label: "Tina Romero" },
-  { value: "Daniel Raim", label: "Daniel Raim" },
-  { value: "Huo Meng", label: "Huo Meng" },
-  { value: "Sylvain Chomet", label: "Sylvain Chomet" },
-  { value: "Cristiano Burlan", label: "Cristiano Burlan" },
-  { value: "Thierry Klifa", label: "Thierry Klifa" },
-  { value: "Urška Djukić", label: "Urška Djukić" },
-  { value: "Francis Ford Coppola", label: "Francis Ford Coppola" },
-  { value: "Fernando Eimbcke", label: "Fernando Eimbcke" },
-  { value: "Max Walker-Silverman", label: "Max Walker-Silverman" },
-  { value: "Tolga Karaçelik", label: "Tolga Karaçelik" },
-  { value: "Dean Francis", label: "Dean Francis" },
-  { value: "Elena Manrique", label: "Elena Manrique" },
-  { value: "Pedro Cabeleira", label: "Pedro Cabeleira" },
-  {
-    value: "Kevin Macdonald e Sam Rice-Edwards",
-    label: "Kevin Macdonald e Sam Rice-Edwards",
-  },
-  {
-    value: "Leela Varghese e Emma Hough Hobb",
-    label: "Leela Varghese e Emma Hough Hobb",
-  },
-  { value: "Kateryna Gornostai", label: "Kateryna Gornostai" },
-  { value: "Oliver Laxe", label: "Oliver Laxe" },
-  { value: "Sven Bresser", label: "Sven Bresser" },
-];
+const props = defineProps({
+  isMenuOpen: { type: Boolean, default: false }
+})
 
-const showcases = [
-  {
-    label: "Première Brasil",
-    value: "premiere-brasil",
-    iconColor: "bg-laranja-600",
-  },
-  {
-    label: "Première Latina",
-    value: "premiere-latina",
-    iconColor: "bg-amarelo-800",
-  },
-];
-
-const isFilterMenuOpen = ref(false);
 const submittedFilters = ref({});
 
-const openMenu = () => {
-  isFilterMenuOpen.value = true;
-  document.body.style.overflow = "hidden";
-};
 const closeMenu = () => {
-  isFilterMenuOpen.value = false;
-  document.body.style.overflow = "";
-};
-
-const searchValue = ref("");
-const cleanInput = () => {
-  searchValue.value = "";
+  emit('close-filter-menu', false)
 };
 
 const timeOptions = computed(() => generateTimeOptions());
@@ -249,7 +43,7 @@ const filters = ref({
   palavrasChaves: null,
 });
 
-const emit = defineEmits(["filtersApplied", "filtersCleared"]);
+const emit = defineEmits(["filtersApplied", "filtersCleared", "close-filter-menu"]);
 
 const applyFilters = async () => {
   const rawFilters = toRaw(filters.value);
@@ -323,57 +117,9 @@ const {
 </script>
 
 <template>
-  <div
-    class="w-full flex flex-col gap-400 md:flex-row md:justify-between md:gap-600"
-  >
-    <div class="input">
-      <div class="relative">
-        <div
-          class="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none"
-        >
-          <IconSearch color="text-primary" />
-        </div>
-        <input
-          v-model="searchValue"
-          type="text"
-          placeholder="Pesquisar"
-          class="w-full md:w-96 pl-10 pr-8 py-2.5 border border-neutrals-300 rounded-[5px] font-body leading-[150%] text-sm text-neutrals-900 placeholder-neutrals-400 focus:outline-none focus:border-neutrals-600 disabled:bg-neutrals-300 disabled:placeholder-neutrals-600 disabled:text-neutrals-600 disabled:border-neutrals-300 disabled:shadow-none transition-all duration-200"
-        />
-        <div class="absolute inset-y-0 right-0 pr-3 flex items-center">
-          <button v-if="searchValue != ''" @click="cleanInput"><IconClose /></button>
-        </div>
-      </div>
-    </div>
-    <div
-      class="filter flex items-center justify-between md:gap-800 lg:gap-1200"
-    >
-      <button
-        @click="openMenu"
-        class="p-100 flex items-center gap-200 text-body-strong-sm text-primary md:order-2"
-      >
-        <IconFilter height="16px" width="16px" color="text-primary" />
-        {{ $t("filter.title") }}
-      </button>
-      <div class="flex items-center gap-300">
-        <span class="text-body-strong-sm uppercase text-secondary-gray"
-          >A - Z</span
-        >
-        <img
-          src="@assets/icons/divisor.svg"
-          alt="divisor"
-          height="16px"
-          width="1px"
-        />
-        <span class="text-body-strong-sm uppercase text-primary">{{
-          $t("filter_by.date")
-        }}</span>
-      </div>
-    </div>
-  </div>
-
   <transition name="slide-left">
     <div
-      v-if="isFilterMenuOpen"
+      v-if="props.isMenuOpen"
       style="margin-top: 0"
       class="fixed inset-0 z-50 bg-white flex flex-col w-full max-w-full h-[100vh] right-0 shadow-lg overflow-y-auto"
     >
